+++ Plugins/UnityPurchasing/Android/IAPResolver/IAPAndroidDependencies.cs
@@ -10,7 +10,6 @@ namespace IAPResolver
         public override List<string> Dependencies =>
             new List<string>()
             {
-                "androidx.activity:activity-compose:1.3.1",
                 "com.android.billingclient:billing:6.2.1"
             };
         public override List<string> Repositories =>